# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardUser = {
  name: string;
  initials: string;
};

export interface StudyTask {
  id: string;
  title: string;
  subject: string;
  duration: number;
  progress: number;
  type: string;
  completed?: boolean;
}

export interface Exam {
  id: string;
  name: string;
  course: string;
  daysLeft: number;
  progress: number;
  date?: string;
}

export interface Dashboard {
  user: DashboardUser;
  todayMinutes: number;
  weeklyMinutes: number;
  streak: number;
  completedTasks: number;
  tasks: StudyTask[];
  exams: Exam[];
}

export interface Course {
  id: string;
  name: string;
  description: string;
  /** @nullable */
  examDate?: string | null;
  documents: number;
  progress: number;
  difficultTopics: number;
  color: string;
}

export interface CourseInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @nullable */
  examDate?: string | null;
}

export interface StudySessionInput {
  courseId: string;
  /** @minimum 1 */
  duration: number;
}

export interface StudySession {
  id: string;
  courseId: string;
  duration: number;
  completedAt: string;
}

export interface Document {
  id: string;
  name: string;
  mime_type: string;
  size_bytes: number;
  status: string;
  created_at?: string;
}

/**
 * @nullable
 */
export type LearningContentSummary = { [key: string]: unknown } | null;

export type LearningContentFlashcardsItem = { [key: string]: unknown };

export type LearningContentQuizItem = { [key: string]: unknown };

export type LearningContentPlanItem = { [key: string]: unknown };

export interface LearningContent {
  /** @nullable */
  summary: LearningContentSummary;
  flashcards: LearningContentFlashcardsItem[];
  quiz: LearningContentQuizItem[];
  plan: LearningContentPlanItem[];
}

export interface StudyPlanItem {
  id: string;
  title: string;
  type?: string;
  duration: number;
  completed: boolean;
  courseId: string;
  scheduledFor?: string;
}