# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: string;
  name: string;
  description: string;
  /** @nullable */
  examDate?: string | null;
  documents: number;
  progress: number;
  difficultTopics: number;
  color: string;
}