# lib/api-zod/src/generated/types/courseInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @nullable */
  examDate?: string | null;
}