# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardUser } from './dashboardUser';
import type { Exam } from './exam';
import type { StudyTask } from './studyTask';

export interface Dashboard {
  user: DashboardUser;
  todayMinutes: number;
  weeklyMinutes: number;
  streak: number;
  completedTasks: number;
  tasks: StudyTask[];
  exams: Exam[];
}