# lib/api-zod/src/generated/types/exam.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Exam {
  id: string;
  name: string;
  course: string;
  daysLeft: number;
  progress: number;
  date?: string;
}