# lib/api-zod/src/generated/types/learningContent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LearningContentFlashcardsItem } from './learningContentFlashcardsItem';
import type { LearningContentPlanItem } from './learningContentPlanItem';
import type { LearningContentQuizItem } from './learningContentQuizItem';
import type { LearningContentSummary } from './learningContentSummary';

export interface LearningContent {
  /** @nullable */
  summary: LearningContentSummary;
  flashcards: LearningContentFlashcardsItem[];
  quiz: LearningContentQuizItem[];
  plan: LearningContentPlanItem[];
}