# lib/api-zod/src/generated/types/studyPlanItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudyPlanItem {
  id: string;
  title: string;
  type?: string;
  duration: number;
  completed: boolean;
  courseId: string;
  scheduledFor?: Date;
}