# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudySession {
  id: string;
  courseId: string;
  duration: number;
  completedAt: string;
}