# lib/api-zod/src/generated/types/studySessionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudySessionInput {
  courseId: string;
  /** @minimum 1 */
  duration: number;
}