# lib/api-zod/src/generated/types/studyTask.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudyTask {
  id: string;
  title: string;
  subject: string;
  duration: number;
  progress: number;
  type: string;
  completed?: boolean;
}